refactor: isolate context-pack status helper
The new read-only handoff-status row introduced pack outcome scanning,
manifest validation, basis freshness checks, and required-role coverage
at the planning selection boundary. Keeping all of that logic inside
artifacts.ts mixed two separate junctions: selecting the approved
planning baseline and classifying context-pack readiness for that
baseline.

Move the pack-specific classifier into context-pack-status.ts and leave
artifacts.ts as the thinner adapter that reads the approved planning
selection and attaches the resulting status fields to approved hints.

Preserve the row surface and behavior:
- artifacts.ts still exports the same status types and helper functions
  added by the row.
- readLatestPlanningArtifacts() and approved hint resolution still
  attach the same contextPack, contextPackStatus,
  missingRequiredContextPackRoles, and contextPackIssues fields.
- The focused planning proof remains unchanged.

This keeps the blast radius closer to the planning helper boundary
without widening the row into consumer gating, runtime projection, or
document refresh work.

Document-refresh: not-needed | current upstream has no upstream-owned
public reference for this planning helper split; proof lives in focused
tests.
